tools/aws/meaocron/requirements.txt
apscheduler
boto3
click
python-decouple
requests
slacker
